+++ src/main/java/net/oneandone/sushi/fs/http/HttpRoot.java
@@ -257,6 +257,11 @@ public class HttpRoot implements Root<HttpNode> {
     //--
 
     @Override
+    public String toString() {
+        return getProtocol() + "://" + getHostname() + ":" + getPort();
+    }
+    
+    @Override
     public boolean equals(Object obj) {
         HttpRoot root;